refactor: move node to node.go
